+++ client-core/src/transaction_builder/default_wallet_transaction_builder.rs
@@ -389,11 +389,7 @@ mod default_wallet_transaction_builder_tests {
                         }
                     }))
                     .unwrap();
-                    // FIXME: all these unit tests don't account for 16-byte tag in encryption
-                    // in those mock encrypt stuff, while the estimation does :/
-                    assert!(
-                        (output_value + fee).unwrap() <= (input_value - Coin::from(16u32)).unwrap()
-                    );
+                    assert!((output_value + fee).unwrap() <= input_value);
 
                     for (i, input) in transaction.inputs.iter().enumerate() {
                         let address = if input.id == [3; 32] {