+++ docker/unittest.sh
@@ -14,10 +14,10 @@ echo "Test $BUILD_MODE $BUILD_PROFILE"
 if [ $BUILD_MODE == "sgx" ]; then
     cargo test $CARGO_ARGS
 else
-    cargo test $CARGO_ARGS --manifest-path client-rpc/Cargo.toml
-    cargo test $CARGO_ARGS --manifest-path client-cli/Cargo.toml
-    cargo test $CARGO_ARGS --features mock-validation --manifest-path dev-utils/Cargo.toml
-    cargo test $CARGO_ARGS --features mock-validation --manifest-path chain-abci/Cargo.toml
+    cargo test $CARGO_ARGS --features mock-enclave --manifest-path client-rpc/Cargo.toml
+    cargo test $CARGO_ARGS --features mock-enclave --manifest-path client-cli/Cargo.toml
+    cargo test $CARGO_ARGS --features mock-enclave --manifest-path dev-utils/Cargo.toml
+    cargo test $CARGO_ARGS --features mock-enclave --manifest-path chain-abci/Cargo.toml
     for pkg in \
         client-common \
         client-network \